merge

+++ core/model/modx/sources/modfilemediasource.class.php
@@ -24,14 +24,14 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
      * @return boolean
      */
     public function initialize() {
-        parent::initialize();
+        $return = parent::initialize();
         $options = array();
         if (!$this->ctx) {
             $this->ctx =& $this->xpdo->context;
         }
         $options['context'] = $this->ctx->get('key');
         $this->fileHandler = $this->xpdo->getService('fileHandler','modFileHandler', '',$options);
-        return true;
+        return $return;
     }
 
     /**
@@ -158,7 +158,7 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
 
                 $dirnames[] = strtoupper($fileName);
                 $directories[$fileName] = array(
-                    'id' => $bases['urlRelative'].rtrim($fileName,'/').'/',
+                    'id' => rawurlencode($bases['urlRelative'].rtrim($fileName,'/').'/'),
                     'text' => $fileName,
                     'cls' => implode(' ',$cls),
                     'iconCls' => 'icon icon-folder',
@@ -190,7 +190,7 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
                 if ($this->hasPermission('file_update') && $canSave) $cls[] = 'pupdate';
 
                 $encFile = rawurlencode($fullPath.$fileName);
-                $page = !empty($editAction) ? '?a='.$editAction.'&file='.$bases['urlRelative'].$fileName.'&wctx='.$this->ctx->get('key').'&source='.$this->get('id') : null;
+                $page = !empty($editAction) ? '?a='.$editAction.'&file='.rawurlencode($bases['urlRelative'].$fileName).'&wctx='.$this->ctx->get('key').'&source='.$this->get('id') : null;
                 $url = $bases['urlRelative'] . $fileName;
 
                 /* get relative url from manager/ */
@@ -199,7 +199,7 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
 
                 $filenames[] = strtoupper($fileName);
                 $files[$fileName] = array(
-                    'id' => $bases['urlRelative'].$fileName,
+                    'id' => rawurlencode($bases['urlRelative'].$fileName),
                     'text' => $fileName,
                     'cls' => implode(' ',$cls),
                     'iconCls' => 'icon icon-file icon-'.$ext . ($file->isWritable() ? '' : ' icon-lock'),
@@ -266,7 +266,7 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
                                     $imageHeight = $imageQueryHeight;
                                 }
                                 $imageQuery = http_build_query(array(
-                                    'src' => $bases['urlRelative'].$fileName,
+                                    'src' => rawurlencode($bases['urlRelative'].$fileName),
                                     'w' => $imageQueryWidth,
                                     'h' => $imageQueryHeight,
                                     'HTTP_MODAUTH' => $modAuth,
@@ -276,7 +276,7 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
                                     'source' => $this->get('id'),
                                     't' => $file->getMTime(),
                                 ));
-                                $image = $this->ctx->getOption('connectors_url', MODX_CONNECTORS_URL).'system/phpthumb.php?'.urldecode($imageQuery);
+                                $image = $this->ctx->getOption('connectors_url', MODX_CONNECTORS_URL).'system/phpthumb.php?'.rawurldecode($imageQuery);
                             } else {
                                 $preview = false;
                                 $this->xpdo->log(modX::LOG_LEVEL_ERROR,'Thumbnail could not be created for file: '.$bases['pathAbsoluteWithPath'].$fileName);
@@ -593,6 +593,25 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
     }
 
     /**
+     * Check that an object (directory, file) exists
+     *
+     * @param string $objectPath The object path to check
+     * @param string $objectName The object name displayed in the error message
+     * @return bool
+     */
+    protected function checkObjectExist($objectPath, $objectName) {
+        if (file_exists($objectPath)) {
+            if (is_dir($objectPath)) {
+                $this->addError('name', $this->xpdo->lexicon('file_folder_err_ae'));
+                return true;
+            }
+            $this->addError('name', sprintf($this->xpdo->lexicon('file_err_ae'), htmlentities($objectName, ENT_QUOTES, 'UTF-8')));
+            return true;
+        }
+        return false;
+    }
+
+    /**
      * @param string $oldPath
      * @param string $newName
      * @return bool
@@ -622,13 +641,7 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
         $newPath = $this->fileHandler->sanitizePath($newName);
         $newPath = dirname($oldPath).'/'.$newPath;
 
-        /* check to see if the new resource already exists */
-        if (file_exists($newPath)) {
-            if (is_dir($newPath)) {
-                $this->addError('name',$this->xpdo->lexicon('file_folder_err_ae'));
-                return false;
-            }
-            $this->addError('name',sprintf($this->xpdo->lexicon('file_err_ae'),$newName));
+        if ($this->checkObjectExist($newPath,$newName)) {
             return false;
         }
 
@@ -696,6 +709,9 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
         $bases = $this->getBases($objectPath);
 
         $fullPath = $bases['pathAbsolute'].$objectPath;
+        if (!$this->checkFiletype($fullPath)) {
+            return false;
+        }
         if (!file_exists($fullPath)) {
             $this->addError('file',$this->xpdo->lexicon('file_folder_err_ns').': '.$fullPath);
             return false;
@@ -744,6 +760,9 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
         $bases = $this->getBases($objectPath);
 
         $fullPath = $bases['pathAbsolute'].ltrim($objectPath,'/');
+        if (!$this->checkFiletype($fullPath)) {
+            return false;
+        }
 
         /** @var modFile $file */
         $file = $this->fileHandler->make($fullPath);
@@ -874,6 +893,11 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
 
             $newPath = $this->fileHandler->sanitizePath($file['name']);
             $newPath = $directory->getPath().$newPath;
+
+            $checkAlreadyExists = (bool)$this->xpdo->getOption('upload_check_exists', null, true);
+            if ($checkAlreadyExists && $this->checkObjectExist($newPath,$file['name'])) {
+                return false;
+            }
 
             if (!move_uploaded_file($file['tmp_name'],$newPath)) {
                 $this->addError('path',$this->xpdo->lexicon('file_err_upload'));
@@ -1053,8 +1077,8 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
                 }
 
                 $filesize = @filesize($filePathName);
-                $url = urlencode(ltrim($dir.$fileName,'/'));
-                $page = !empty($editAction) ? '?a='.$editAction.'&file='.$bases['urlRelative'].$fileName.'&wctx='.$this->ctx->get('key').'&source='.$this->get('id') : null;
+                $url = rawurlencode(ltrim($dir.$fileName,'/'));
+                $page = !empty($editAction) ? '?a='.$editAction.'&file='.rawurlencode($bases['urlRelative'].$fileName).'&wctx='.$this->ctx->get('key').'&source='.$this->get('id') : null;
 
                 /* get thumbnail */
                 $preview = 0;
@@ -1096,7 +1120,7 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
                             $thumbHeight = $size[1] >= $thumbHeight ? $thumbHeight : $size[1];
                             $thumbWidth = round($size[0] * ($thumbHeight / $size[1]));
                         }
-                        $image = $thumb = $bases['urlAbsolute'].urldecode($url);
+                        $image = $thumb = $bases['urlAbsolute'].rawurldecode($url);
                     } else {
                         $size = @getimagesize($filePathName);
                         if (is_array($size) && $size[0] > 0 && $size[1] > 0) {
@@ -1133,7 +1157,7 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
                                 'source' => $this->get('id'),
                                 't' => $file->getMTime(),
                             ));
-                            $image = $this->ctx->getOption('connectors_url', MODX_CONNECTORS_URL).'system/phpthumb.php?'.urldecode($imageQuery);
+                            $image = $this->ctx->getOption('connectors_url', MODX_CONNECTORS_URL).'system/phpthumb.php?'.rawurldecode($imageQuery);
                             $thumbQuery = http_build_query(array(
                                 'src' => $url,
                                 'w' => $thumbQueryWidth,
@@ -1145,7 +1169,7 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
                                 'source' => $this->get('id'),
                                 't' => $file->getMTime(),
                             ));
-                            $thumb = $this->ctx->getOption('connectors_url', MODX_CONNECTORS_URL).'system/phpthumb.php?'.urldecode($thumbQuery);
+                            $thumb = $this->ctx->getOption('connectors_url', MODX_CONNECTORS_URL).'system/phpthumb.php?'.rawurldecode($thumbQuery);
                         } else {
                             $this->xpdo->log(modX::LOG_LEVEL_ERROR,'Thumbnail could not be created for file: '.$filePathName);
                             $preview = 0;
@@ -1162,7 +1186,7 @@ class modFileMediaSource extends modMediaSource implements modMediaSourceInterfa
 
                 $filenames[] = strtoupper($fileName);
                 $files[$fileName] = array(
-                    'id' => $bases['urlAbsoluteWithPath'].$fileName,
+                    'id' => rawurlencode($bases['urlAbsoluteWithPath'].$fileName),
                     'name' => $fileName,
                     'cls' => 'icon-'.$fileExtension,
                     'image' => $image,
